WaterConservancyIoT/requirements.txt
# 智能水利系统 - Python依赖包列表
# 生成时间: 2025-09-14

# 核心框架和服务器
redis==5.0.1
mysql-connector-python==8.2.0
websockets==12.0

# 异步支持
aioredis==2.0.1
asyncio

# 安全和认证
bcrypt==4.1.1
PyJWT==2.8.0

# 数据处理
json
datetime
typing
logging

# 系统工具
random
time
threading
urllib.parse
os

# 开发和调试 (可选)
# pytest==7.4.3
# pytest-asyncio==0.21.1
